# projects/mark-ms/lib/swagger.yaml
...
paths:
...
  /account/login:
    post:
      summary: Allows a user to log into the Mark system
      description: Allows a user to log into the Mark system.
      tags:
        - User Accounts
      responses:
        '200':
          $ref: '#/responses/200 Successful Authorization'
          
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Authorization'
          
        - $ref: '#/parameters/Handle Body'
          
        - $ref: '#/parameters/Key String'
          
    x-oad-type: operation
  /account/check-handle-availability:
    post:
      summary: Check is user handle is available.
      description: Check to see if a user handle is available.
      tags:
        - User Accounts
      responses:
        '200':
          description: Handle is available
          schema:
            x-oad-type: 'null'
            type: 'null'
          x-oad-type: response
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Handle Body'
          
    x-oad-type: operation
  /account/signup:
    post:
      summary: Start of the signup process.
      description: >-
        The start of the signup process.Returns a code, roll, and hashed key.
        Perform another hash on the hashed key and return the roll and the
        state, along with the texted code in the signup-validate step.
      tags:
        - User Accounts
      responses:
        '200':
          description: Successful signup attempt.
          schema:
            x-oad-type: object
            type: object
            properties:
              State:
                x-oad-type: string
                type: string
                title: State
                description: State of transaction.
              roll:
                x-oad-type: integer
                type: integer
                title: Roll
                description: Roll number
                format: int32
              key:
                x-oad-type: string
                type: string
                title: key
                description: A hashed key.
          x-oad-type: response
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/State'
          
        - $ref: '#/parameters/Roll'
          
        - $ref: '#/parameters/Key'
          
    x-oad-type: operation
  /account/signup-validate:
    post:
      summary: Validate the user hardware identity
      description: >-
        Validate that a user is in posession of the phone which they claim to
        own.
      tags:
        - User Accounts
      responses:
        '200':
          description: Successful identity verification.
          schema:
            x-oad-type: object
            type: object
            title: token
            description: >-
              A user token that will allow the user to create/modify content
              created by the respective token owner on the platform.
            properties:
              token:
                x-oad-type: string
                type: string
                title: token
                description: >-
                  A user token that will allow the user to create/modify content
                  created by the respective token owner on the platform.
          x-oad-type: response
        '400':
          $ref: '#/responses/400 Bad Request'
          
        '403':
          $ref: '#/responses/403 Forbidden'
          
        '422':
          $ref: '#/responses/422 Unprocessable Entity'
          
        '429':
          $ref: '#/responses/429 Too Many Requests'
          
        '500':
          $ref: '#/responses/500 Server Error'
          
      parameters:
        - $ref: '#/parameters/Key'
          
        - $ref: '#/parameters/Roll'
          
        - $ref: '#/parameters/State'
          
        - $ref: '#/parameters/Code'
          
    x-oad-type: operation
...
parameters:
...
  Authorization:
    name: Authorization
    in: header
    description: User access token.
    required: true
    type: string
    x-oad-type: parameter
...
  Key:
    name: key
    in: body
    required: true
    schema:
      x-oad-type: string
      type: string
      title: key
    x-oad-type: parameter
  Roll:
    name: roll
    in: body
    required: true
    schema:
      x-oad-type: integer
      type: integer
      title: roll
      format: int32
    x-oad-type: parameter
  State:
    name: state
    in: body
    required: true
    schema:
      x-oad-type: integer
      type: integer
      title: state
      format: int32
    x-oad-type: parameter
  Code:
    name: code
    in: body
    required: true
    schema:
      x-oad-type: integer
      type: integer
      format: int32
    x-oad-type: parameter
...
securityDefinitions:
  Authorization:
    description: User authentication token
    name: User Token
    in: header
    type: apiKey
info:
  title: Mark Services API
  version: 1.0.0
  description: Communicate with the Mark Services project
  contact:
    name: Jake Ferrante
    email: ferrantej@knights.ucf.edu
  license:
    name: MIT
    url: 'https://mit-license.org'
host: mark-services.azurewebsites.net
basePath: /api/v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
swagger: '2.0'